Extracted unconfigured repo dropdown to separate module
