+++ src/common/components/repository-row/dropdowns/index.js
@@ -1,7 +1,9 @@
 import NameMismatchDropdown from './name-mismatch-dropdown';
 import RemoveRepoDropdown from './remove-repo-dropdown';
+import UnconfiguredDropdown from './unconfigured-dropdown';
 
 export {
   NameMismatchDropdown,
-  RemoveRepoDropdown
+  RemoveRepoDropdown,
+  UnconfiguredDropdown
 };
